Runs EMtree for several covariates choices

Runs EMtree for several covariates choices

ComparEMtree(
  counts,
  covar_matrix,
  models,
  m_names,
  O = NULL,
  unlinked = NULL,
  Pt = 0.1,
  v = 0.8,
  S = 100,
  maxIter = 50,
  cond.tol = 1e-14,
  cores = 3
)

Arguments

counts
Data of observed counts with dimensions n x p, either a matrix, data.frame or tibble.

covar_matrix
matrix of covariates, should have the same number of rows as the count matrix.

models
list of covariate combinations to be tested. For example list(1,2) will design two linear models with the first two covariates adjusted separately

m_names
list of names for the models to be compared, for example list("first model","last model")

O
Offset matrix with dimensions n x p

unlinked
An optional vector of nodes which are not linked with each other

Pt
Probability threshold for every sub-sample

v
The proportion of observed data to be taken in each sub-sample. It is the ratio (sub-sample size)/n

S
Number of desired sub-samples

maxIter
Maximum number of iterations of EMtree

cond.tol
Tolerance for the psi matrix.

cores
Number of cores, can be greater than 1 if data involves less than about 32 species.

Value

a tibble in a suitable format for graphical purposes. It has four columns describing each edge of each model: the endpoints ("node1", "node2"), the model it corresponds to ("model") and its weight ("value").

Examples

n=30
p=10
S=3
Y = data_from_scratch(type="tree",p=p,n=n)$data
X = data.frame(rnorm(n),rbinom(n,1,0.7))
ComparEMtree(Y,X,models=list(1,2,c(1,2)),m_names=list("1","2","both"),Pt=0.3,S=S, cores=1)

#> model 1 : Computing  3 probability matrices with 1 core(s)... 
#> Convergence took 0.08 secs  and  8  iterations.
#> Convergence took 0.21 secs  and  12  iterations.
#> Convergence took 0.09 secs  and  9  iterations.0.43 secs 
#> 
#> model 2 : Computing  3 probability matrices with 1 core(s)... 
#> Convergence took 0.1 secs  and  10  iterations.
#> Convergence took 0.11 secs  and  10  iterations.
#> Convergence took 0.09 secs  and  10  iterations.0.32 secs 
#> 
#> model both : Computing  3 probability matrices with 1 core(s)... 
#> Convergence took 0.09 secs  and  13  iterations.
#> Convergence took 0.08 secs  and  9  iterations.
#> Convergence took 0.07 secs  and  8  iterations.0.26 secs 
#> 

#> # A tibble: 135 x 4
#>    node1 node2 model weight
#>    <chr> <chr> <chr>  <dbl>
#>  1 1     2     1          0
#>  2 1     3     1          1
#>  3 2     3     1          1
#>  4 1     4     1          0
#>  5 2     4     1          0
#>  6 3     4     1          0
#>  7 1     5     1          0
#>  8 2     5     1          0
#>  9 3     5     1          0
#> 10 4     5     1          1
#> # … with 125 more rows