Resampling procedure for edges probability

Resampling procedure for edges probability

ResampleEMtree(counts, covar_matrix = NULL, O = NULL, v = 0.8,
  S = 100, maxIter = 30, cond.tol = 1e-14, cores = 3)

Arguments

counts
Data of observed counts with dimensions n x p, either a matrix, data.frame or tibble.

covar_matrix
matrix of covariates, should have the same number of rows as the count matrix.

O
Matrix of offsets, with dimension n x p

v
The proportion of observed data to be taken in each sub-sample. It is the ratio (sub-sample size)/n

S
Total number of wanted sub-samples.

maxIter
Maximum number of EMtree iterations at each sub-sampling.

cond.tol
Tolerance for the psi matrix.

cores
Number of cores, can be greater than 1 if data involves less than about 32 species.

Value

Returns a list which contains the Pmat data.frame, and vectors of EMtree maximum iterations and running times in each resampling.

Pmat: S x p(p-1)/2 matrix with edge probabilities for each resample

maxIter: EMtree maximum iterations in each resampling.

times: EMtree running times in each resampling.

Examples

n=30
p=10
S=5
Y=data_from_scratch("tree",p=p,n=n)$data
X = data.frame(rnorm(n),rbinom(n,1,0.7))
resample=ResampleEMtree(Y,covar_matrix=X, S=S,cores = 1)

#> 
#> S= 1   
#> S= 2   
#> S= 3   
#> S= 4   
#> S= 5   

str(resample)

#> List of 3
#>  $ Pmat   : num [1:5, 1:45] 0.595 0.835 0.921 0.86 0.704 ...
#>  $ maxIter: num [1:5] 3 3 3 3 3
#>  $ times  : 'difftime' num [1:5] 0.0468077659606934 0.0512855052947998 0.0477886199951172 0.0579118728637695 ...
#>   ..- attr(*, "units")= chr "secs"