generates data under the PLN model with a certain type of dependency structure, and draws the structure.

generates data under the PLN model with a certain type of dependency structure, and draws the structure.

data_from_scratch(type, p = 20, n = 50, r = 5, covariates = NULL,
  prob = log(p)/p, dens = log(p)/p, draw = FALSE)

Arguments

type
type of graph, either "tree", "erdos", "cluster" or "scale-free"

p
wanted number of columns (species)

n
wanted number of rows (samples/observations)

r
within/between connectiviy ratio for cluster graphs

covariates
a data.frame or matrix containing data covariates.

prob
edge probability for erdos graphs

dens
density of edges for cluster graphs

draw
boolean, plots the graph if set to TRUE

Value

a list containing

data: simulated counts

omega: the precision matrix

Examples

data_from_scratch("tree",p=10,draw=TRUE)

#> [1] 10 10

#> $data
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    2    3    1    1    0    1    1    1    1     1
#>  [2,]    3    3    2    3    0    0    0    1    2     8
#>  [3,]    0    0    1    0    1    5    5    0    0     1
#>  [4,]    2    1    0    0    1    0    2    3    1     1
#>  [5,]    1    0    5    1    3    2   13    8    1     1
#>  [6,]    2    5    0    3    0    0    0    1    4     2
#>  [7,]    1    1    0    0    1    0    0    0    0     1
#>  [8,]    2    0    6    0    7    3    2    0    0     0
#>  [9,]    5    4    0    0    0    0    0    1    6     0
#> [10,]    4    2    0    1    0    0    0    2    2     4
#> [11,]    0    1    0    1    0    0    0    0    1     0
#> [12,]    8    1    0    5    0    0    1    0    4     9
#> [13,]    2    1    0    5    0    0    0    0    2     1
#> [14,]    0    0    2    2    1    0    3    2    2     0
#> [15,]    0    1    1    0    0    2    2    0    0     1
#> [16,]   12   10    1   16    0    0    0    0   16     8
#> [17,]    0    0    4    1    1    3    4    2    1     0
#> [18,]    1    2    1    2    1    1    0    0    0     1
#> [19,]    2    2    0    1    0    0    0    1    0     0
#> [20,]    0    1    8    1    7    7    2    2    0     0
#> [21,]    0    0    7    0    6    6   12   14    0     0
#> [22,]    0    0    0    1    2    2    4    2    0     0
#> [23,]    0    1    1    1    1    3    1    2    2     1
#> [24,]    0    0    2    0    0    0    4    6    1     0
#> [25,]    4    0    3    0    3    1    2    0    2     0
#> [26,]    2    3    0    1    0    0    0    1    3     1
#> [27,]    5    1    0    8    0    0    0    0    8    11
#> [28,]    2    3    0    4    0    1    0    2    3     3
#> [29,]    9    4    0    8    0    0    0    0    4     7
#> [30,]    3    3    1    1    0    1    1    1    0     1
#> [31,]    3    1    0    0    0    0    4    0    1     2
#> [32,]    1    0    0    3    1    0    1    0    2     1
#> [33,]    4    0    0    1    2    1    1    2    0     2
#> [34,]    1    1    2    0    4    0    1    0    1     2
#> [35,]    0    0   14    0    7    8    5    3    0     0
#> [36,]    1    1    0    1    1    0    0    5    0     1
#> [37,]    3    0    0    2    0    0    0    1    3     1
#> [38,]    2    1    0    1    0    0    2    0    2     0
#> [39,]    1    0    2    1    2    2    4    4    0     1
#> [40,]    0    0    7    0    5    8    1   14    0     0
#> [41,]    3    0    1    2    2    0    0    0    2     5
#> [42,]    2    0    1    1    1    0    2    3    3     0
#> [43,]    2    2    0    2    1    0    1    0    4     2
#> [44,]    0    0    0    2    1    1    2    3    0     0
#> [45,]    2    6    0    1    0    0    1    3    2     1
#> [46,]    2    2    1    0    0    0    1    0    2     3
#> [47,]    0    2    3    1    0    1    0    0    0     1
#> [48,]    0    1    1    0    1    1    3    2    0     0
#> [49,]    2    0    1    1    1    3    7    3    0     1
#> [50,]    0    1    0    1    0    3    0    3    1     0
#> 
#> $omega
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,] 1.01 0.00 0.00 0.00 0.00 1.00 0.00 0.00 0.00  0.00
#>  [2,] 0.00 2.01 0.00 0.00 1.00 0.00 1.00 0.00 0.00  0.00
#>  [3,] 0.00 0.00 1.01 0.00 0.00 0.00 0.00 0.00 0.00  1.00
#>  [4,] 0.00 0.00 0.00 2.01 1.00 0.00 0.00 1.00 0.00  0.00
#>  [5,] 0.00 1.00 0.00 1.00 3.01 0.00 0.00 0.00 1.00  0.00
#>  [6,] 1.00 0.00 0.00 0.00 0.00 2.01 0.00 0.00 0.00  1.00
#>  [7,] 0.00 1.00 0.00 0.00 0.00 0.00 1.01 0.00 0.00  0.00
#>  [8,] 0.00 0.00 0.00 1.00 0.00 0.00 0.00 2.01 0.00  1.00
#>  [9,] 0.00 0.00 0.00 0.00 1.00 0.00 0.00 0.00 1.01  0.00
#> [10,] 0.00 0.00 1.00 0.00 0.00 1.00 0.00 1.00 0.00  3.01
#>