Generate simulation parameters from a graph adjacency matrix

Generate simulation parameters from a graph adjacency matrix

generator_param(G, signed = FALSE, v = 0.01)

Arguments

G
adjacency matrix

signed
boolean: should the graph be composed of positive and negative partial correlations ?

v
parameter controlling the noise on the precision matrix

Value

a list containing

sigma: the covariance matrix

omega: the precision matrix

cste: the constant that was needed to load the diagonal of omega and esure its positive definitiveness

Examples

G=generator_graph(p=10,graph="tree")
generator_param(G=G)

#> $sigma
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,]  1.0000000  0.9131742 -0.7946788 -0.7946788  0.9113284 -0.8693083
#>  [2,]  0.9131742  1.0000000 -0.7950154 -0.7950154  0.9117144 -0.8696766
#>  [3,] -0.7946788 -0.7950154  1.0000000  0.9162207 -0.8720005  0.9141507
#>  [4,] -0.7946788 -0.7950154  0.9162207  1.0000000 -0.8720005  0.9141507
#>  [5,]  0.9113284  0.9117144 -0.8720005 -0.8720005  1.0000000 -0.9538914
#>  [6,] -0.8693083 -0.8696766  0.9141507  0.9141507 -0.9538914  1.0000000
#>  [7,]  0.8302169  0.8305686 -0.9571942 -0.9571942  0.9109965 -0.9550316
#>  [8,] -0.9553992 -0.9558039  0.8317767  0.8317767 -0.9538718  0.9098902
#>  [9,] -0.9578977 -0.8747275  0.7612210  0.7612210 -0.8729593  0.8327084
#> [10,] -0.8704011 -0.8707698  0.8328395  0.8328395 -0.9550906  0.9110527
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,]  0.8302169 -0.9553992 -0.9578977 -0.8704011
#>  [2,]  0.8305686 -0.9558039 -0.8747275 -0.8707698
#>  [3,] -0.9571942  0.8317767  0.7612210  0.8328395
#>  [4,] -0.9571942  0.8317767  0.7612210  0.8328395
#>  [5,]  0.9109965 -0.9538718 -0.8729593 -0.9550906
#>  [6,] -0.9550316  0.9098902  0.8327084  0.9110527
#>  [7,]  1.0000000 -0.8689739 -0.7952629 -0.8700841
#>  [8,] -0.8689739  1.0000000  0.9151746  0.9110340
#>  [9,] -0.7952629  0.9151746  1.0000000  0.8337552
#> [10,] -0.8700841  0.9110340  0.8337552  1.0000000
#> 
#> $omega
#> 10 x 10 sparse Matrix of class "dsCMatrix"
#>                                                        
#>  [1,] 2.01 .    .    .    .    .    .    1.00 1.00 .   
#>  [2,] .    1.01 .    .    .    .    .    1.00 .    .   
#>  [3,] .    .    1.01 .    .    .    1.00 .    .    .   
#>  [4,] .    .    .    1.01 .    .    1.00 .    .    .   
#>  [5,] .    .    .    .    3.01 1.00 .    1.00 .    1.00
#>  [6,] .    .    .    .    1.00 2.01 1.00 .    .    .   
#>  [7,] .    .    1.00 1.00 .    1.00 3.01 .    .    .   
#>  [8,] 1.00 1.00 .    .    1.00 .    .    3.01 .    .   
#>  [9,] 1.00 .    .    .    .    .    .    .    1.01 .   
#> [10,] .    .    .    .    1.00 .    .    .    .    1.01
#> 
#> $cste
#> [1] 1
#>