Runs EMtree for several covariates choices

Runs EMtree for several covariates choices

ComparEMtree(counts, covar_matrix, models, m_names, O = NULL, Pt = 0.1,
  v = 0.8, S = 100, maxIter = 50, cond.tol = 1e-14, cores = 3)

Arguments

counts
Data of observed counts with dimensions n x p, either a matrix, data.frame or tibble.

covar_matrix
matrix of covariates, should have the same number of rows as the count matrix.

models
list of covariate combinations to be tested. For example list(1,2) will design two linear models with the first two covariates adjusted separately

m_names
list of names for the models to be compared, for example list("first model","last model")

O
Offset matrix with dimensions n x p

Pt
Probability threshold for every sub-sample

v
The proportion of observed data to be taken in each sub-sample. It is the ratio (sub-sample size)/n

S
Number of desired sub-samples

maxIter
Maximum number of iterations of EMtree

cond.tol
Tolerance for the psi matrix.

cores
Number of cores, can be greater than 1 if data involves less than about 32 species.

Value

a tibble in a suitable format for graphical purposes. It has four columns describing each edge of each model: the origin node ("node1"), the destination node ("node2"), the model it corresponds to ("model") and its weight ("value").

Examples

n=30
p=10
S=3
Y = data_from_scratch(type="tree",p=p,n=n)$data
X = data.frame(rnorm(n),rbinom(n,1,0.7))
ComparEMtree(Y,X,models=list(1,2,c(1,2)),m_names=list("1","2","both"),Pt=0.3,S=S, cores=1)

#> 
#> model  1
#> S= 1  [1] 0.9583333
#> 
#> Convergence took 1.21 secs  and  3  iterations.  0.9583333
#> S= 2  [1] 0.9583333
#> 
#> Convergence took 0.05 secs  and  3  iterations.  0.9583333
#> S= 3  [1] 0.9583333
#> 
#> Convergence took 0.05 secs  and  3  iterations.  0.9583333
#> model  2
#> S= 1  [1] 0.9583333
#> 
#> Convergence took 0.08 secs  and  6  iterations.  0.9583333
#> S= 2  [1] 0.9583333
#> 
#> Convergence took 0.06 secs  and  3  iterations.  0.9583333
#> S= 3  [1] 0.9583333
#> 
#> Convergence took 0.09 secs  and  5  iterations.  0.9583333
#> model  both
#> S= 1  [1] 0.9583333
#> 
#> Convergence took 0.09 secs  and  3  iterations.  0.9583333
#> S= 2  [1] 0.9583333
#> 
#> Convergence took 0.05 secs  and  3  iterations.  0.9583333
#> S= 3  [1] 0.9583333
#> 
#> Convergence took 0.05 secs  and  3  iterations.  0.9583333

#> # A tibble: 135 x 4
#>    node1 node2 model weight
#>    <chr> <chr> <chr>  <dbl>
#>  1 1     2     1      0    
#>  2 1     3     1      1    
#>  3 2     3     1      0    
#>  4 1     4     1      0.333
#>  5 2     4     1      0.333
#>  6 3     4     1      1    
#>  7 1     5     1      0    
#>  8 2     5     1      0    
#>  9 3     5     1      0.333
#> 10 4     5     1      0    
#> # … with 125 more rows