Generate simulation parameters from a graph adjacency matrix

Generate simulation parameters from a graph adjacency matrix

generator_param(G, signed = FALSE, v = 0.01)

Arguments

G
adjacency matrix

signed
boolean: should the graph be composed of positive and negative partial correlations ?

v
parameter controlling the noise on the precision matrix

Value

a list containing

sigma: the covariance matrix

omega: the precision matrix

cste: the constant that was needed to load the diagonal of omega and esure its positive definitiveness

Examples

G=generator_graph(p=10,graph="tree")
generator_param(G=G)

#> $sigma
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,]  10.553731 -10.071131 -10.347799  -9.593309  10.245346  10.245346
#>  [2,] -10.071131  10.564871   9.874616  10.063624  -9.776848  -9.776848
#>  [3,] -10.347799   9.874616  11.126373   9.406118 -11.016211 -10.045432
#>  [4,]  -9.593309  10.063624   9.406118  11.519382  -9.312988  -9.312988
#>  [5,]  10.245346  -9.776848 -11.016211  -9.312988  11.897239   9.945972
#>  [6,]  10.245346  -9.776848 -10.045432  -9.312988   9.945972  11.897239
#>  [7,] -10.347799   9.874616  10.145886   9.406118 -10.045432 -11.016211
#>  [8,]   9.689242 -10.164261  -9.500179 -10.634576   9.406118   9.406118
#>  [9,]  -9.500179   9.965929   9.314805  10.427067  -9.222580  -9.222580
#> [10,]   9.406118  -9.867256  -9.222580 -10.323829   9.131267   9.131267
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,] -10.347799   9.689242  -9.500179   9.406118
#>  [2,]   9.874616 -10.164261   9.965929  -9.867256
#>  [3,]  10.145886  -9.500179   9.314805  -9.222580
#>  [4,]   9.406118 -10.634576  10.427067 -10.323829
#>  [5,] -10.045432   9.406118  -9.222580   9.131267
#>  [6,] -11.016211   9.406118  -9.222580   9.131267
#>  [7,]  11.126373  -9.500179   9.314805  -9.222580
#>  [8,]  -9.500179  10.740922 -10.531338  10.427067
#>  [9,]   9.314805 -10.531338  11.306330 -11.194387
#> [10,]  -9.222580  10.427067 -11.194387  12.073650
#> 
#> $omega
#> 10 x 10 sparse Matrix of class "dsCMatrix"
#>                                                        
#>  [1,] 3.01 1.00 1.00 .    .    .    1.00 .    .    .   
#>  [2,] 1.00 2.01 .    .    .    .    .    1.00 .    .   
#>  [3,] 1.00 .    2.01 .    1.00 .    .    .    .    .   
#>  [4,] .    .    .    1.01 .    .    .    1.00 .    .   
#>  [5,] .    .    1.00 .    1.01 .    .    .    .    .   
#>  [6,] .    .    .    .    .    1.01 1.00 .    .    .   
#>  [7,] 1.00 .    .    .    .    1.00 2.01 .    .    .   
#>  [8,] .    1.00 .    1.00 .    .    .    3.01 1.00 .   
#>  [9,] .    .    .    .    .    .    .    1.00 2.01 1.00
#> [10,] .    .    .    .    .    .    .    .    1.00 1.01
#> 
#> $cste
#> [1] 1
#>