Generate simulation parameters from a graph adjacency matrix

Generate simulation parameters from a graph adjacency matrix

generator_param(G, signed = FALSE, v = 0.01)

Arguments

G
adjacency matrix

signed
boolean: should the graph be composed of positive and negative partial correlations ?

v
parameter controlling the noise on the precision matrix

Value

a list containing

sigma: the covariance matrix

omega: the precision matrix

cste: the constant that was needed to load the diagonal of omega and esure its positive definitiveness

Examples

G=generator_graph(p=10,graph="tree")
generator_param(G=G)

#> $sigma
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,]  11.168473   9.510747  -9.605855   9.797021  -9.700021  -9.979767
#>  [2,]   9.510747  11.736302 -10.853665  10.079564  -9.979767  -9.325167
#>  [3,]  -9.605855 -10.853665  10.962201 -10.180360  10.079564   9.418419
#>  [4,]   9.797021  10.079564 -10.180360  10.382959 -10.280158  -9.605855
#>  [5,]  -9.700021  -9.979767  10.079564 -10.280158  11.168473   9.510747
#>  [6,]  -9.979767  -9.325167   9.418419  -9.605855   9.510747  11.736302
#>  [7,]  10.079564   9.418419  -9.512603   9.701913  -9.605855 -10.853665
#>  [8,] -10.280158  -9.605855   9.701913  -9.894991   9.797021  10.079564
#>  [9,]  -9.700021  -9.979767  10.079564 -10.280158  10.178374   9.510747
#> [10,]  10.178374   9.510747  -9.605855   9.797021  -9.700021  -9.979767
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,]  10.079564 -10.280158  -9.700021  10.178374
#>  [2,]   9.418419  -9.605855  -9.979767   9.510747
#>  [3,]  -9.512603   9.701913  10.079564  -9.605855
#>  [4,]   9.701913  -9.894991 -10.280158   9.797021
#>  [5,]  -9.605855   9.797021  10.178374  -9.700021
#>  [6,] -10.853665  10.079564   9.510747  -9.979767
#>  [7,]  10.962201 -10.180360  -9.605855  10.079564
#>  [8,] -10.180360  10.382959   9.797021 -10.280158
#>  [9,]  -9.605855   9.797021  11.168473  -9.700021
#> [10,]  10.079564 -10.280158  -9.700021  11.168473
#> 
#> $omega
#> 10 x 10 sparse Matrix of class "dsCMatrix"
#>                                                        
#>  [1,] 1.01 .    .    .    .    .    .    1.00 .    .   
#>  [2,] .    1.01 1.00 .    .    .    .    .    .    .   
#>  [3,] .    1.00 2.01 1.00 .    .    .    .    .    .   
#>  [4,] .    .    1.00 4.01 1.00 .    .    1.00 1.00 .   
#>  [5,] .    .    .    1.00 1.01 .    .    .    .    .   
#>  [6,] .    .    .    .    .    1.01 1.00 .    .    .   
#>  [7,] .    .    .    .    .    1.00 2.01 1.00 .    .   
#>  [8,] 1.00 .    .    1.00 .    .    1.00 4.01 .    1.00
#>  [9,] .    .    .    1.00 .    .    .    .    1.01 .   
#> [10,] .    .    .    .    .    .    .    1.00 .    1.01
#> 
#> $cste
#> [1] 1
#>