Generate simulation parameters from a graph adjacency matrix

Generate simulation parameters from a graph adjacency matrix

generator_param(G, signed = FALSE, v = 0.01)

Arguments

G
adjacency matrix

signed
boolean: should the graph be composed of positive and negative partial correlations ?

v
parameter controlling the noise on the precision matrix

Value

a list containing

sigma: the covariance matrix

omega: the precision matrix

cste: the constant that was needed to load the diagonal of omega and esure its positive definitiveness

Examples

G=generator_graph(p=10,graph="tree")
generator_param(G=G)

#> $sigma
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,]  1.0000000  0.9108039 -0.9551529  0.9129019 -0.9535579  0.9129019
#>  [2,]  0.9108039  1.0000000 -0.8699570  0.8314746 -0.9551638  0.8314746
#>  [3,] -0.9551529 -0.8699570  1.0000000 -0.8719609  0.9107935 -0.8719609
#>  [4,]  0.9129019  0.8314746 -0.8719609  1.0000000 -0.8705048  0.9150575
#>  [5,] -0.9535579 -0.9551638  0.9107935 -0.8705048  1.0000000 -0.8705048
#>  [6,]  0.9129019  0.8314746 -0.8719609  0.9150575 -0.8705048  1.0000000
#>  [7,] -0.9543329 -0.8692102  0.9115339 -0.9565864  0.9100117 -0.9565864
#>  [8,]  0.8341742  0.8355790 -0.7967639  0.7615192 -0.8748018  0.7615192
#>  [9,] -0.8701924 -0.8716580  0.8311668 -0.7944003  0.9125743 -0.7944003
#> [10,]  0.9100337  0.9115663 -0.8692213  0.8307715 -0.9543560  0.8307715
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,] -0.9543329  0.8341742 -0.8701924  0.9100337
#>  [2,] -0.8692102  0.8355790 -0.8716580  0.9115663
#>  [3,]  0.9115339 -0.7967639  0.8311668 -0.8692213
#>  [4,] -0.9565864  0.7615192 -0.7944003  0.8307715
#>  [5,]  0.9100117 -0.8748018  0.9125743 -0.9543560
#>  [6,] -0.9565864  0.7615192 -0.7944003  0.8307715
#>  [7,]  1.0000000 -0.7960799  0.8304533 -0.8684751
#>  [8,] -0.7960799  1.0000000 -0.9586089  0.9166410
#>  [9,]  0.8304533 -0.9586089  1.0000000 -0.9562201
#> [10,] -0.8684751  0.9166410 -0.9562201  1.0000000
#> 
#> $omega
#> 10 x 10 sparse Matrix of class "dsCMatrix"
#>                                                        
#>  [1,] 3.01 .    1.00 .    1.00 .    1.00 .    .    .   
#>  [2,] .    1.01 .    .    1.00 .    .    .    .    .   
#>  [3,] 1.00 .    1.01 .    .    .    .    .    .    .   
#>  [4,] .    .    .    1.01 .    .    1.00 .    .    .   
#>  [5,] 1.00 1.00 .    .    3.01 .    .    .    .    1.00
#>  [6,] .    .    .    .    .    1.01 1.00 .    .    .   
#>  [7,] 1.00 .    .    1.00 .    1.00 3.01 .    .    .   
#>  [8,] .    .    .    .    .    .    .    1.01 1.00 .   
#>  [9,] .    .    .    .    .    .    .    1.00 2.01 1.00
#> [10,] .    .    .    .    1.00 .    .    .    1.00 2.01
#> 
#> $cste
#> [1] 1
#>